core: extract world filesystem helpers
